+++ run_tests.py
@@ -213,9 +213,10 @@ def main():
 
     our_version = stem.version.get_system_tor_version(args.tor_path)
     skipped_targets = []
+    integ_setup_thread = None
 
     if not args.specific_test or 'test.integ.installation'.startswith(args.specific_test):
-      test.integ.installation.setup()
+      integ_setup_thread = test.integ.installation.setup()
 
     for target in args.run_targets:
       # check if we meet this target's tor version prerequisites
@@ -247,6 +248,9 @@ def main():
         # We should have joined on all threads. If not then that indicates a
         # leak that could both likely be a bug and disrupt further targets.
 
+        if integ_setup_thread:
+          integ_setup_thread.join()
+
         active_threads = threading.enumerate()
 
         if len(active_threads) > 1:
@@ -267,6 +271,9 @@ def main():
       except OSError:
         error_tracker.register_error()
       finally:
+        if integ_setup_thread:
+          test.integ.installation.clean()
+
         println()
         integ_runner.stop()
         println()